+++ src/test/groovy/ru/vyarus/dropwizard/guice/config/support/conf/ConfigLevel1.groovy
@@ -1,0 +1,10 @@
+package ru.vyarus.dropwizard.guice.config.support.conf
+
+import io.dropwizard.Configuration
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 19.06.2016
+ */
+abstract class ConfigLevel1 extends Configuration implements Level1Interface, Cloneable {
+}