+++ src/test/groovy/ru/vyarus/dropwizard/guice/config/support/conf/ConfigLevel2.groovy
@@ -1,0 +1,8 @@
+package ru.vyarus.dropwizard.guice.config.support.conf
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 19.06.2016
+ */
+class ConfigLevel2 extends ConfigLevel1 implements Level2Interface, Serializable {
+}