+++ src/test/groovy/ru/vyarus/dropwizard/guice/config/support/conf/Level1Interface.groovy
@@ -1,0 +1,8 @@
+package ru.vyarus.dropwizard.guice.config.support.conf
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 19.06.2016
+ */
+interface Level1Interface extends Level1IndirectInterface {
+}