+++ src/test/groovy/ru/vyarus/dropwizard/guice/config/support/conf/Level2Interface.groovy
@@ -1,0 +1,9 @@
+package ru.vyarus.dropwizard.guice.config.support.conf
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 19.06.2016
+ */
+interface Level2Interface {
+
+}